Removed events, direction and status from API.
All tests pass.  Here is a summary of the impact of each item removed:

o Events - the evolution of exchange API, and the getMessage() method
  in particular, have made this part of the API pretty redundant.
  I still like the idea of being notified of events related to the
  exchange, but I think we should wait until we have a clear cut
  case for including them.

o Direction - this was originally included under the assumption that
  a handler could be used in consuming and providing roles.  In this
  case, you would need to know the direction of the exchange to
  determine whether you wanted to respond to the event.  Simplifying
  things here by removing Direction and treating the single handler
  case as an edge case.  You can check out MockHandler in the test
  module to see how the single handler implementation will look.

o Status - this was tied to events, i.e. which event got triggered
  was based on the status of the exchange.  Now that events are gone,
  we don't need status.

One thing that's missing at present is the ability to stop processing
of an event in the handler chain without throwing an exception.  I
have some ideas on how I want to handle this and will commit something
later today.