SWITCHYARD-181 Initial implementation of HornetQ Dispatch provider.
The initial implementation creates an embedded instance of a HornetQ
server per provider instance.  See javadoc in HornetQBus for information
related to configuration of the provider and HornetQ.

Exchange serialization has not been implemented at this point, so the
initial implementation has a hacked serialiazation to just move bits
of the exchange for testing purposes.  Once serialization is complete,
this implementation will be updated.