SWITCHYARD-111 workaround for inability to resolve wsdl20-instance.xsd
This is a temporary fix only.  I have removed the include for sca-binding-ws-1.1-cd04-rev2.xsd, which imports wsdl20-instance.xsd for the wsdlLocation attribute.  I'm not sure why this schema fails to resolve while all the other schema work fine.  This needs to be investigated and fixed, at which point this workaround can be removed.